A mistake! Please undo if possible.
You can select the Skycam in play mode with number button 2 and scroll down. At the end of the scroll, the camera pans slightly. Scripts are poorly written; Functions are bad; Control via the Inspector too disorganized. The only thing I like here is the pan at the end.
The other scripts are also useless - we can certainly do better ;-)

Maybe we can test how the reset works and what it looks like on GitHub.